Move the preview traits and wasm URL prelude to i-slint-live-preview
LspToPreview and PreviewToLsp move next to the message enums they wrap,
under the protocol feature; the unused std::any::Any supertrait is
dropped. The protocol module also gains the Error/Result aliases the
traits need and the wasm-only UrlWasm prelude that previously lived in
the LSP's wasm crate root. tools/lsp re-exports everything through
common, so consumer paths are unchanged.
