+++ internal/live-preview/Cargo.toml
@@ -20,6 +20,7 @@ lsp-types = { workspace = true, optional = true }
 serde = { workspace = true, optional = true }
 serde_json = { workspace = true, optional = true }
 derive_more = { workspace = true, features = ["debug"], optional = true }
+tracing = { workspace = true, optional = true }
 # file-watcher
 notify = { version = "8.0.0", default-features = false, features = ["macos_kqueue"], optional = true }
 i-slint-compiler = { workspace = true, optional = true }
@@ -34,7 +35,6 @@ futures-util = { version = "0.3.31", features = ["sink"], optional = true }
 postcard = { version = "1.1.3", features = ["alloc"], optional = true }
 base64 = { version = "0.22.1", optional = true }
 hostname = { version = "0.4.1", optional = true }
-tracing = { workspace = true, optional = true }
 anyhow = { version = "1.0.100", optional = true }
 
 [target.'cfg(not(any(target_vendor = "apple", target_arch = "wasm32")))'.dependencies]
@@ -47,7 +47,7 @@ getifs = { version = "0.6.1", optional = true }
 path = "lib.rs"
 
 [features]
-protocol = ["dep:lsp-types", "dep:serde", "dep:serde_json", "dep:derive_more"]
+protocol = ["dep:lsp-types", "dep:serde", "dep:serde_json", "dep:derive_more", "dep:tracing"]
 file-watcher = ["dep:notify", "dep:i-slint-compiler"]
 live-component = ["file-watcher", "dep:slint-interpreter", "dep:i-slint-core"]
 ffi = ["live-component", "slint-interpreter/ffi"]
@@ -56,7 +56,6 @@ remote = [
   "dep:slint-interpreter",
   "dep:i-slint-core",
   "dep:i-slint-compiler",
-  "dep:tracing",
   "dep:dashmap",
   "dep:tokio-tungstenite",
   "dep:tokio",
