+++ tools/lsp/preview/drop_location.rs
@@ -20,7 +20,7 @@ use i_slint_live_preview::protocol::VersionedUrl;
 use crate::preview::ext::ElementRcNodeExt;
 
 #[cfg(target_arch = "wasm32")]
-use crate::wasm_prelude::*;
+use crate::common::wasm_prelude::*;
 
 pub fn placeholder() -> String {
     format!(
