+++ tools/lsp/util.rs
@@ -15,7 +15,7 @@ use smol_str::SmolStr;
 use crate::common;
 
 #[cfg(target_arch = "wasm32")]
-use crate::wasm_prelude::UrlWasm;
+use crate::common::wasm_prelude::UrlWasm;
 
 /// Get the `TextRange` of a `node`, excluding any trailing whitespace tokens.
 pub fn node_range_without_trailing_ws(node: &SyntaxNode) -> TextRange {
