+++ tools/lsp/wasm_main.rs
@@ -29,27 +29,9 @@ use std::rc::Rc;
 use wasm_bindgen::prelude::*;
 
 #[cfg(target_arch = "wasm32")]
-use crate::wasm_prelude::*;
+use crate::common::wasm_prelude::*;
 
 type JsResult<T> = std::result::Result<T, JsError>;
-
-pub mod wasm_prelude {
-    use std::path::{Path, PathBuf};
-
-    /// lsp_url doesn't have method to convert to and from PathBuf for wasm, so just make some
-    pub trait UrlWasm {
-        fn to_file_path(&self) -> Result<PathBuf, ()>;
-        fn from_file_path<P: AsRef<Path>>(path: P) -> Result<lsp_types::Url, ()>;
-    }
-    impl UrlWasm for lsp_types::Url {
-        fn to_file_path(&self) -> Result<PathBuf, ()> {
-            Ok(self.to_string().into())
-        }
-        fn from_file_path<P: AsRef<Path>>(path: P) -> Result<Self, ()> {
-            Self::parse(path.as_ref().to_str().ok_or(())?).map_err(|_| ())
-        }
-    }
-}
 
 impl RequestHandler {
     fn handle_request(
